Decrypt mode

+++ RedesII-TII/Decrypt/Model/Keys.cs
@@ -8,13 +8,13 @@ namespace RedesII_TII.Model
 {
     public class PublicKey
     {
-        public int n { get; set; }
+        public ushort n { get; set; }
         public int d { get; set; }
     }
 
     public class PrivateKey
     {
-        public int n { get; set; }
+        public ushort n { get; set; }
         public int e { get; set; }
     }
